[mdatagen] fix missing/extraneous imports for telemetry codegen
the telemetry code generation does not include some required imports
and also includes some imports that are not used, which leads to build failures.
this fixes those issues; re-running mdatagen on the samplereceiver results
in a no-op, indicating that somehow there was drift b/w the output codegen
and the templates.