[chore] add text about unmaintained vendor components (#11616)
#### Description

This PR adds language to our definition of `unmaintained` that allows
vendor-specific components to become unmaintained if they lose an active
code owner from the contributing vendor. This is necessary to prevent a
vendor-specific component form being maintained only by the community
instead of by the vendor. Since vendors had privileged access to getting
components accepted in contrib they must be held to a higher standard
for maintaining those components.

---------

Co-authored-by: Alex Boten <223565+codeboten@users.noreply.github.com>
Co-authored-by: Pablo Baeyens <pbaeyens31+github@gmail.com>