[chore] Add internal attribute package (#12073)
This PR creates a new internal package which defines the correct set of
attributes for each kind of component.

This is a subset of #12057 which is broken off in order to reduce the
overall size of that PR. As such, this package will not actually be used
until #12057 is merged.