[chore] [builder] add metadata.yaml (#12592)
<!--Ex. Fixing a bug - Describe the bug and how this fixes the issue.
Ex. Adding a feature - Explain what this achieves.-->
#### Description
This PR adds a `metadata.yaml` file to `cmd/builder` to allow it to show
up in the issue templates. This makes it so the component is part of the
component dropdown.

<!-- Issue number if applicable -->
#### Link to tracking issue
I didn't open an issue cause it's just a chore task, but I can open one
if needed.

<!--Describe what testing was performed and which tests were added.-->
#### Testing
Not applicable.

<!--Describe the documentation added.-->
#### Documentation
It's in the issue templates now!

<!--Please delete paragraphs that you did not use before submitting.-->