[chore][ping-codeowners-on-new-issue] Ping even if owner filed (#12716)
<!--Ex. Fixing a bug - Describe the bug and how this fixes the issue.
Ex. Adding a feature - Explain what this achieves.-->
#### Description
Code owners should be pinged even if a code owner filed an issue. This
helps other code owners, in the case of more than one for a single
component, be notified of relevant issues being filed.

This is the same change as
https://github.com/open-telemetry/opentelemetry-collector-contrib/pull/38841,
as
[requested](https://github.com/open-telemetry/opentelemetry-collector-contrib/pull/38841#issuecomment-2747326094).