update semconv dependency to otelgo's semconv package
This updates the calls to the collector's internal semconv package with
otelgo's instead. The main difference is how the keys can be used, they
will need to be cast as strings where they key name is needed, otherwise
it's not a huge change. This would allow us to stop producing our own
semconv package once contrib is moved as well.

Signed-off-by: Alex Boten <223565+codeboten@users.noreply.github.com>