[confighttp,configgrpc] Rename TLSSetting to TLS (#13115)
<!--Ex. Fixing a bug - Describe the bug and how this fixes the issue.
Ex. Adding a feature - Explain what this achieves.-->
#### Description

Renames `TLSSetting` fields to `TLS`. The name is a historical artifact
since the `configtls` had `Setting` in the name before. This is more
consistent with the rest.

I think if we do this change we should do it alongside other changes to
confighttp, to avoid breaking our users multiple times.