[processor/batchprocessor] Improve batchprocessor worst case perf
As the batch size approaches one, the time complexity of the batch
processor approaches O(n^2). This is caused by repeatedly counting the
number of data points in a message. This commit introduces a
"short-circuit" behavior that allows the core loop to exit quickly when
the desired number of data points has been found.