telemetry: enable adding attributes from context
Add a ContextWithAttributes function to the component
package, which will add an attribute.Set to the given
context.

The context attributes will be added to any spans and
synchronous metric measurements made within the context.

The context attributes are not added to any log records,
as the Zap logger API does not support contextual logging.