service: refactor logger wrapping
Preparation work for https://github.com/open-telemetry/opentelemetry-collector/issues/4970

There are two main changes here:
 - We move the componentattribute logger core wrapping and Logger to LoggerProvider
   tee'ing to the service package. This will enable the wrapping to be performed
   independently of the implementation of telemetry providers.
 - The LoggerProvider is now wrapped such that it filters out log records by
   severity, according to the configured level.

The second change is necessary due to the first: because the LoggerProvider will
in the future be injectable, and because the telemetry configuration will be opaque
to the service package, we must always tee the logs from the zap logger to the
LoggerProvider. When a tee core is created, the resulting level of the core will
be the lowest level of either core. Moreover, the LoggerProvider may be entirely
disabled, in which case the existing use of zapcore.NewIncreaseLevelCore is unsafe
and may panic.

There's also some light refactoring of tests, and the service package
tests are now ~30s faster by replacing a sleep with assert.Eventually.