[cmd/mdatagen] add func to configure attribute set
Adding the ability to automatically add attributes via an attribute set when
recording telemetry through the TelemetryBuilder for internal metrics. This wraps
the otel go SDK functions for convenience. I have a follow up PR that demonstrates
how to use this, we *could* hide the otel go metrics directly, but that can be
done in a follow up PR if it's useful.

Signed-off-by: alex boten <223565+codeboten@users.noreply.github.com>