[chore] update readme badges to align the sizes (#13352)
<!--Ex. Fixing a bug - Describe the bug and how this fixes the issue.
Ex. Adding a feature - Explain what this achieves.-->
#### Description
The badges currently in main branch:

![image](https://github.com/user-attachments/assets/98486920-7ca2-40a2-b98a-f92b180dea88)

The badges after the fix:

![image](https://github.com/user-attachments/assets/0ea50453-942d-440e-a650-a8672504af01)
Now the smaller ones are at the same line, makes it looks a little bit
better.

Sorry this may looks like a spam PR. I created it as I previously can
not make my PR go through the merge queue in this repo, as I am still
interested in some work in this repo, I raised this pr to test if my
commits / CLA can pass the merge queue now.


<!-- Issue number if applicable -->
#### Link to tracking issue
Fixes #

<!--Describe what testing was performed and which tests were added.-->
#### Testing

<!--Describe the documentation added.-->
#### Documentation

<!--Please delete paragraphs that you did not use before submitting.-->