Introduce service.Settings.TelemetryFactory
The `service.Settings` type now has a `TelemetryFactory`
field for injecting the `telemetry.Factory` to be used
for creating a logger and logger provider, meter provider,
and tracer provider.

The `otelcol` package is hard-coded to inject an
otelconftelemetry factory for now. In a followup we will
make it possible to inject the telemetry through
`otelcol.Factories`.