Make telemetry register with zpages
Instead of having zpagesextension register a zpages
span processor on the tracer provider, make this the
responsibility of the telemetry implementation. To make
this happen, the service now creates the zpages mux and
exposes it as a host capability. Components may register
their own zpages routes, which the zpagesextension does
for exposing expvar; and otherwise the zpagesextension
just exists to expose the HTTP endpoint.