Add ReusePort config to confighttp
This adds in a new field, `ReusePort` that, if set, sets the
SO_REUSEPORT socket option on the listener port. If we're on non unix,
this errors out instead as AFAIK SO_REUSEPORT isn't available. Cursory
testing says that SO_REUSEADDR _might_ work, but I don't have a windows
machine to test on.

Signed-off-by: sinkingpoint <colin@quirl.co.nz>