change AsPartial to IsPartial
The linter did not like exporting a private type, which is a fair point.
I want to keep the error type private so that the only way to produce a
partial error also necessitates that it's permanent, so I changed the
API to `IsPartial` which produces a count and a boolean.