Fix prepare release 0.140.0 again (#14173)
#### Description

It seems `internal/testutil` was added under `excluded-modules:` in
`versions.yaml`, which prevents it from being versioned. This is
probably a mistake, since the module is imported by tests in various
other modules.

This PR makes `internal/testutil` a regular beta module, and updates
imports to it to 0.140.0 since multimod did not do it the first time
around.